% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmax-package.R
\docType{package}
\name{pharmax-package}
\alias{pharmax}
\alias{pharmax-package}
\title{pharmax: Public Package Coordinator}
\description{
Coordinates the current public pharmax package surface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pharmax-org/pharmax}
  \item \url{https://pharmax-org.github.io/pharmax}
  \item Report bugs at \url{https://github.com/pharmax-org/pharmax/issues}
}

}
\author{
\strong{Maintainer}: Malek Okour \email{malek@zajaly.com}

}
\keyword{internal}
