% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_status.R
\name{px_public_packages}
\alias{px_public_packages}
\title{List public pharmax packages}
\usage{
px_public_packages()
}
\value{
A data frame with package names, roles, and public status labels.
}
\description{
List public pharmax packages
}
